Organizing the files Rayan and Ricardo
