--- src/WishListRepository.java
@@ -1,4 +1,0 @@
-@Repository
-public interface WishListRepository extends JpaRepository<WishList, Long> {
-    List<WishList> findByUserId(Long userId);
-}
