+++ src/main/java/book/ratings/project/repository/BookRepository.java
@@ -3,6 +3,8 @@ import book.ratings.project.model.Book;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 import java.util.Optional;
 
 
@@ -14,7 +16,7 @@ public interface BookRepository  extends JpaRepository<Book,Long> {
     Optional<Book> findByAuthor(String author);
 
 
+    List<Book> findByRatingGreaterThanEqual(double rating);
 
-
-
+    List<Book> findTop10ByOrderByCopiesSoldDesc();
 }
