+++ src/main/java/books/Book.java
@@ -1,10 +1,10 @@
-package geekText;
+package books;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "books") // Annotation from the master branch
@@ -59,5 +59,14 @@ public class Book {
 
     public void setRating(double rating) {
         this.rating = rating;
+    }
+
+    public void setPublisher(String publisher) {
+    }
+
+    public void setYearPublished(int yearPublished) {
+    }
+
+    public void setRatings(double ratings) {
     }
 }
