+++ src/main/java/books/BookController.java
@@ -1,9 +1,11 @@
-package geekText;
+package books;
 
-import com.example.books.entity.Book;
-import com.example.books.repository.BookRepository;
+//import com.example.books.entity.books.Book;
+//import com.example.books.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import book.ratings.project.repository.BookRepository;
+import book.ratings.project.model.Book;
 
 import java.util.List;
 
@@ -19,7 +21,7 @@ public class BookController {
     public List<Book> getTopSellers() {
         return bookRepository.findTop10ByOrderByCopiesSoldDesc();
 
-        public List<Book> getBooksByRating(@RequestParam double rating) {
-            return bookRepository.findByRatingGreaterThanEqual(rating);
+//        public List<Book> getBooksByRating(@RequestParam double rating) {
+//            return bookRepository.findByRatingGreaterThanEqual(rating);
     }
 }
