+++ src/main/java/books/Entity.java
@@ -1,4 +1,4 @@
-package geekText;
+package books;
 
 public @interface Entity {
 }
