--- src/main/java/org/example/Book.java
@@ -1,65 +1,0 @@
-package geekText;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Book {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String title;
-    private String author;
-    private int yearPublished;
-    private String genre;
-    private String publisher;
-    private double ratings;
-
-    // Getters and setters
-    public Long getId() {
-        return id;
-    }
-    public void setId(Long id) {
-        this.id = id;
-    }
-    public String getTitle() {
-        return title;
-    }
-    public void setTitle(String title) {
-        this.title = title;
-    }
-    public String getAuthor() {
-        return author;
-    }
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-    public int getYearPublished() {
-        return yearPublished;
-    }
-    public void setYearPublished(int yearPublished) {
-        this.yearPublished = yearPublished;
-    }
-    public String getGenre() {
-        return genre;
-    }
-    public void setGenre(String genre) {
-        this.genre = genre;
-    }
-    public String getPublisher() {
-        return publisher;
-    }
-    public void setPublisher(String publisher) {
-        this.publisher = publisher;
-    }
-    public double getRatings() {
-        return ratings;
-    }
-    public void setRatings(double ratings) {
-        this.ratings = ratings;
-    }
-}
