--- src/main/java/org/example/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package geekText;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
