+++ src/main/java/wishlist/Main.java
@@ -1,3 +1,5 @@
+package wishlist;
+
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
 public class Main {
