+++ src/main/java/wishlist/WishListRepository.java
@@ -1,9 +1,11 @@
-package geekText;
+package wishlist;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
+@Repository
 public interface WishListRepository extends JpaRepository<WishList, Long> {
     List<WishList> findByUserId(Long userId);
-    boolean existsByUserIdAndName(Long userId, String name);
 }
